requirements.txt
librosa>=0.10.0
soundfile>=0.12.0
scipy>=1.10.0
numpy>=1.24.0
vocos>=0.1.0
torch>=2.0.0
phonemizer>=3.2.0
inflect>=7.0.0
dnsmos>=0.1.0
pyyaml>=6.0
tqdm>=4.65.0
